fix handling of odd lengths in swab function
this function is specified to leave the last byte with "unspecified
disposition" when the length is odd, so for the most part correct
programs should not be calling swab with odd lengths. however, doing
so is permitted, and should not write past the end of the destination
buffer.

(cherry picked from commit dccbf4c809efc311aa37da71de70d04dfd8b0db3)